+++ src/core/entity-manager.js
@@ -14,6 +14,7 @@ class EntityManager {
         this.eventHandler     = new EventHandler()
         
         this.entityConfigurations = new Map()
+        this.componentLookup      = new Map()
         
         this.entities = Array.from({ length : this.capacity }, () => ({ components: 0 }))
     }
@@ -26,13 +27,31 @@ class EntityManager {
         this.entities = [...this.entities, ...Array.from({ length : oldCapacity }, () => ({ components: 0 }))]
 
         for (let i = oldCapacity; i < this.capacity; ++i) {
-            for (let component of this.componentManager.getComponents().keys()) {
-                this.entities[i][component] = this.componentManager.newComponent(component)
+            let entity = this.entities[i]
+            
+            for (const componentId of this.componentManager.getComponents().keys()) {
+                let componentName = null
+                
+                for (let [key, value] of this.componentLookup.entries()) {
+                    if (value === componentId) {
+                        componentName = key
+                        
+                        break
+                    }
+                }
+
+                entity[componentId] = this.componentManager.newComponent(componentId)
+                
+                Object.defineProperty(entity, componentName, { get() { return this[componentId] }})
             }
         }
     }
     
     newEntity(components) {
+        if (Array.isArray(components)) {
+            components = Array.from(this.componentLookup).reduce((curr, next) => ['', curr[1] | next[1]], ['', 0])[1]
+        }
+        
         if (!Number.isInteger(components) || components <= 0) {
             return { id : this.capacity, entity : null }
         }
@@ -78,6 +97,8 @@ class EntityManager {
         this.currentMaxEntity = 0
     }
 
+    // Does not allow components to be anything other than a bitmask for performance reasons
+    // This method will be called for every system for every loop (which might be as high as 60 / second)
     *getEntities(components = 0) {
         for (let id = 0; id <= this.currentMaxEntity; ++id) {
             if (components === 0 || (this.entities[id].components & components) === components) {
@@ -96,11 +117,18 @@ class EntityManager {
     
     // Component Manager
     
-    registerComponent(component) {
+    registerComponent(name, component) {
+        if (typeof name !== 'string' || name.length === 0) {
+            throw TypeError('name must be a non-empty string.')
+        }
+        
         const componentId = this.componentManager.registerComponent(component)
+        
+        this.componentLookup.set(name, componentId)
         
         for (let entity of this.entities) {
             entity[componentId] = this.componentManager.newComponent(componentId)
+            Object.defineProperty(entity, name, { get() { return this[componentId] }})
         }
         
         let initializer
@@ -124,30 +152,42 @@ class EntityManager {
         return componentId
     }
     
-    addComponent(entityId, componentId) {
-        this.entities[entityId].components |= componentId
+    addComponent(entityId, component) {
+        if (typeof component === 'string') {
+            this.entities[entityId].components |= this.componentLookup.get(component)
+        } else {
+            this.entities[entityId].components |= component
+        }
     }
     
-    removeComponent(entityId, componentId) {
-        this.entities[entityId].components &= ~componentId
+    removeComponent(entityId, component) {
+        if (typeof component === 'string') {
+            this.entities[entityId].components &= ~this.componentLookup.get(component)
+        } else {
+            this.entities[entityId].components &= ~component   
+        }
     }
     
     // System Manager
     
     registerSystem(type, components, callback) {
+        if (Array.isArray(components)) {
+            components = Array.from(this.componentLookup).reduce((curr, next) => ['', curr[1] | next[1]], ['', 0])[1]
+        }
+        
         return this.systemManager.registerSystem(type, components, callback)
     }
     
     registerLogicSystem(components, callback) {
-        return this.systemManager.registerSystem(SystemType.Logic, components, callback)
+        return this.registerSystem(SystemType.Logic, components, callback)
     }
     
     registerRenderSystem(components, callback) {
-        return this.systemManager.registerSystem(SystemType.Render, components, callback)
+        return this.registerSystem(SystemType.Render, components, callback)
     }
     
     registerInitSystem(components, callback) {
-        return this.systemManager.registerSystem(SystemType.Init, components, callback)
+        return this.registerSystem(SystemType.Init, components, callback)
     }
     
     removeSystem(systemId) {
@@ -174,8 +214,12 @@ class EntityManager {
     
     // Entity Factory
     
-    registerInitializer(componentId, initializer) {
-        this.entityFactory.registerInitializer(componentId, initializer)
+    registerInitializer(component, initializer) {
+        if (typeof component === 'string') {
+            this.entityFactory.registerInitializer(this.componentLookup.get(component), initializer)
+        } else {
+            this.entityFactory.registerInitializer(component, initializer)
+        }
     }
     
     build() {
@@ -184,8 +228,12 @@ class EntityManager {
         return this
     }
     
-    withComponent(componentId, initializer) {
-        this.entityFactory.withComponent(componentId, initializer)
+    withComponent(component, initializer) {
+        if (typeof component === 'string') {
+            this.entityFactory.withComponent(this.componentLookup.get(component), initializer)
+        } else {
+            this.entityFactory.withComponent(component, initializer)
+        }
         
         return this
     }