+++ test/entity-manager/increase-capacity.js
@@ -7,18 +7,24 @@ describe('EntityManager', function() {
             this.entityManager = new EntityManager(100)
             
             this.position = 1
+            this.positionName = 'position'
             this.positionComponent = { x : 1, y : 1, z : -2 }
             
             this.velocity = 2
+            this.velocityName = 'velocity'
             this.velocityComponent = 0.25
             
             this.stats = 4
+            this.statsName = 'stats'
             this.statsComponent = { xp : 10000, level : 20 }
             
-            this.components = this.position | this.velocity | this.stats
             this.entityManager.componentManager.components.set(this.position, this.positionComponent)
             this.entityManager.componentManager.components.set(this.velocity, this.velocityComponent)
             this.entityManager.componentManager.components.set(this.stats, this.statsComponent)
+            
+            this.entityManager.componentLookup.set(this.positionName, this.position)
+            this.entityManager.componentLookup.set(this.velocityName, this.velocity)
+            this.entityManager.componentLookup.set(this.statsName, this.stats)
         })
         
         afterEach(() => {
@@ -82,6 +88,24 @@ describe('EntityManager', function() {
                 expect(this.entityManager.entities[i][this.position]).to.deep.equal(this.positionComponent)
                 expect(this.entityManager.entities[i][this.velocity]).to.deep.equal(this.velocityComponent)
                 expect(this.entityManager.entities[i][this.stats]).to.deep.equal(this.statsComponent)
+            }
+        })
+        
+        it('adds component getters (by name) to the newly created entities as the capacity is increased', () => {
+            this.entityManager.increaseCapacity()
+            
+            for (let i = 100; i < this.entityManager.capacity; ++i) {
+                expect(this.entityManager.entities[i][this.positionName]).to.deep.equal(this.positionComponent)
+                expect(this.entityManager.entities[i][this.velocityName]).to.deep.equal(this.velocityComponent)
+                expect(this.entityManager.entities[i][this.statsName]).to.deep.equal(this.statsComponent)
+            }
+            
+            this.entityManager.increaseCapacity()
+            
+            for (let i = 200; i < this.entityManager.capacity; ++i) {
+                expect(this.entityManager.entities[i][this.positionName]).to.deep.equal(this.positionComponent)
+                expect(this.entityManager.entities[i][this.velocityName]).to.deep.equal(this.velocityComponent)
+                expect(this.entityManager.entities[i][this.statsName]).to.deep.equal(this.statsComponent)
             }
         })
     })